Update Fibonacci_numbers.cpp
Поправил типы данных, для большей вместительности

+++ 00_Funny/Fibonacci_numbers.cpp
@@ -4,15 +4,15 @@ int main()
 {
     int n{};
     std::cin >> n;
-    int x = 1;
-    int y = 0;
+    long long x = 1;
+    long long y = 0;
     for (int i = 0; i < n; i++)
     {
         std::cout << x << ' ';
-        int z = x;
+        long long z = x;
         x = x + y;
         y = z;
     }
     return 0;
 }
-// Просто смешная фигня, выводит последовательность из n чисел Фибоначчи
+// Просто прикольный скрипт, выводит последовательность из n чисел Фибоначчи
